split model-viewer chunk
